+++ src/lib/kotlin/slatekit-entities/src/main/kotlin/slatekit/entities/Entities.kt
@@ -21,8 +21,8 @@ import slatekit.common.log.LogsDefault
 import slatekit.common.naming.Namer
 import slatekit.common.utils.ListMap
 import slatekit.entities.core.*
-import slatekit.entities.repos.EntityMapperEmpty
-import slatekit.entities.repos.InMemoryRepo
+import slatekit.entities.EntityMapperEmpty
+import slatekit.entities.EntityRepoInMemory
 import slatekit.data.support.IdGenerator
 import slatekit.data.support.LongIdGenerator
 import slatekit.meta.models.Model
@@ -168,7 +168,7 @@ open class Entities(
 
         // 3. Repo ( provides CRUD using the Mapper)
         val info = EntityInfo(entityIdType, entityType, table, '`', model, this.enc, this.namer)
-        val repo = InMemoryRepo<TId, T>(info, entityIdGen)
+        val repo = EntityRepoInMemory<TId, T>(info, entityIdGen)
 
         // 4. Service ( used to provide validation, placeholder for business functionality )
         val service = builder.service(this, serviceType, repo, serviceCtx)
